KVM: VMX: Add __alloc_vmcs_cpu() for VMXON region allocation
Split the common logic of allocating VMCS in alloc_vmcs_cpu() as
__alloc_vmcs_cpu() and use the later for VMXON region allocation to
simplify the eVMCS handling when allocating the VMXON regions.

For now KVM uses a common alloc_vmcs_cpu() to allocate a VMCS for all
three cases: 1) VMXON region; 2) a normal VMCS to run a vcpu; 3) a
shadow VMCS.  For a normal "hyperv-enlightened VMCS" (eVMCS) running in
L1, its "revision_id" needs to be set to a "hyperv-defined" value rather
than the value reported by the hardware.  But a VMXON region in L1 when
eVMCS is enabled still needs the hardware value.

Currently KVM simply uses alloc_vmcs_cpu(), which sets "revision_id" to
the "hyperv-defined" value when eVMCS is enabled, to allocate the VMXON
region, and then immediately resets the "revision_id" to the hardware's
value.  It's unnecessary.

Introduce __alloc_vmcs_cpu() to allocate a VMCS and always set
"revision_id" to the hardware value, and use it for VMXON region
allocation to avoid such unnessarily step.

Signed-off-by: Kai Huang <kai.huang@intel.com>
Signed-off-by: Chao Gao <chao.gao@intel.com>
Reviewed-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>