KVM: Disable virtualization in common code on emergency reboot
Virtualization disabling should be done in the common code rather than
a user of it since doing that in a user may surprise other users. Also
set kvm_rebooting in the common code so that every user can observe this
no matter whether its notifier is called before or after kvm's.

Signed-off-by: Chao Gao <chao.gao@intel.com>