KVM: Make virtualization enabling built-in
Make virtualization enabling built-in as it no longer depends on KVM.
Expose necessary functions for KVM and other components to call.

Rename major and exported functions to remove KVM from their names.
Ideally, we should rename all functions and move related files out of
virt/kvm and x86/kvm. The effort required is non-trivial. So, leave this as
a future work.

Signed-off-by: Chao Gao <chao.gao@intel.com>