KVM: x86: Refactor VMX enabling/disabling into a common file
Move VMX enaling/disabling logic out of kvm-intel into a common file.
So, VMX can be enabled from a common place for other purposes like
TDX.

PT handling on VMX enabling/disabling is also moved because this
handling is a hardware requirement rather than a KVM thing.

Unexport vmxarea as it is not used in other places.

Signed-off-by: Chao Gao <chao.gao@intel.com>